moved dependencies to useRef and some components to seperate files to remove lint-warnings
