+++ src/components/Content/PlotWidget/PlotSettingsPopup/PlotSettingsPopup.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import {
     Dialog,
     DialogTitle,
@@ -77,10 +77,12 @@ const PlotSettingsPopup: React.FC<PlotSettingsPopupProps> = ({
         )
     );
     const [localPlotSettings, setLocalPlotSettings] = useState(plotSettings);
+    const plotSettingsRef = useRef(plotSettings);
 
     // When the widget is opened, update the local Plot Settings
     useEffect(() => {
         if (open) {
+            const plotSettings = plotSettingsRef.current;
             setLocalPlotSettings(plotSettings);
 
             setInputsMin(
@@ -94,7 +96,7 @@ const PlotSettingsPopup: React.FC<PlotSettingsPopupProps> = ({
                 )
             );
         }
-    }, [open, plotSettings]);
+    }, [open]);
 
     // For the sake of simplicity, there is no safe button, updates are applied on safe.
     const saveAndClose = useCallback(() => {
