+++ src/components/Content/PlotWidget/PlotWidget.tsx
@@ -20,7 +20,7 @@ import {
     Y_AXIS_ASSIGNMENT_OPTIONS,
     USED_Y_AXES,
 } from "./PlotWidget.types";
-import { useApiUrls } from "../../ApiContext/ApiContext";
+import { useApiUrls } from "../../ApiContext/ApiContextHooks";
 import axios, { AxiosError, AxiosResponse } from "axios";
 import { cloneDeep, isEqual } from "lodash";
 import * as styles from "./PlotWidget.styles";
@@ -288,7 +288,12 @@ const PlotWidget: React.FC<PlotWidgetProps> = React.memo(
             [backendUrl]
         );
 
+        const onUpdatePlotSettingsRef = useRef(onUpdatePlotSettings);
+        const curveAttributesRef = useRef(curveAttributes);
+        const yAxisAttributesRef = useRef(yAxisAttributes);
+
         useEffect(() => {
+            const onUpdatePlotSettings = onUpdatePlotSettingsRef.current;
             onUpdatePlotSettings(index, {
                 plotTitle: plotTitle,
                 curveAttributes: curveAttributes,
@@ -301,10 +306,12 @@ const PlotWidget: React.FC<PlotWidgetProps> = React.memo(
             curveAttributes,
             yAxisAttributes,
             manualAxisAssignment,
-            onUpdatePlotSettings,
         ]);
 
         useEffect(() => {
+            const curveAttributes = curveAttributesRef.current;
+            const yAxisAttributes = yAxisAttributesRef.current;
+
             const newAxisOptions = Y_AXIS_ASSIGNMENT_OPTIONS;
             const newCurveAttributes = new Map<string, CurveAttributes>();
             const newYAxisAttributes = new Array(...yAxisAttributes);
@@ -389,12 +396,7 @@ const PlotWidget: React.FC<PlotWidgetProps> = React.memo(
             ) {
                 setCurveAttributes(newCurveAttributes);
             }
-        }, [
-            channels,
-            getChannelIdentifier,
-            manualAxisAssignment,
-            yAxisAttributes,
-        ]);
+        }, [channels, getChannelIdentifier, manualAxisAssignment]);
 
         useEffect(() => {
             const handleKeyDown = (event: KeyboardEvent) => {
@@ -1922,8 +1924,14 @@ const PlotWidget: React.FC<PlotWidgetProps> = React.memo(
             [backendUrl]
         );
 
+        const handleDoubleClickRef = useRef(handleDoubleClick);
+        const layoutRef = useRef(layout);
+
         useEffect(() => {
             const currentPlotDiv = plotRef.current;
+            const handleDoubleClick = handleDoubleClickRef.current;
+            const layout = layoutRef.current;
+
             if (currentPlotDiv) {
                 plotlyDataRef.current = cloneDeep(data);
                 plotlyConfigRef.current = cloneDeep(config);
