+++ src/components/Content/PlotWidget/WaveformPreviewPopup/WaveformPreviewPopup.tsx
@@ -478,14 +478,17 @@ const WaveformPreviewPopup: React.FC<WaveformPreviewPopupProps> = ({
         [yAxisTimestamps, visibleTimestamps]
     );
 
+    const layoutRef = useRef(layout);
+
     useEffect(() => {
         const currentPlotDiv = plotRef.current;
+        const layout = layoutRef.current;
         if (currentPlotDiv) {
             Plotly.react(currentPlotDiv, data, layout || {}, {
                 displaylogo: false,
             });
         }
-    }, [data, layout]);
+    }, [data]);
 
     useEffect(() => {
         const currentPlotDiv = plotRef.current;
